refactored to make project tidier going foward

+++ src/lib.rs
@@ -2,12 +2,12 @@ use itertools::Itertools;
 
 pub mod input_data;
 
-pub fn find_2020_pair(vals_in: (usize, Vec<i64>) ) -> i64 {
+pub fn find_2020_pair(vals_in: (usize, Vec<usize>) ) -> usize {
     let vals = vals_in.1;
     let mut value = 0;
     let val_pairs = vals.into_iter().combinations(vals_in.0);
     for i in val_pairs.into_iter() {
-        if i.iter().sum::<i64>() == 2020 {
+        if i.iter().sum::<usize>() == 2020 {
             value = i.iter().product();
         }
     }
